*Add observer to banner.tsx
*remove observer from posts
after rebase
